functions/hello/requirements.txt
crowdstrike-foundry-function==1.1.1
